Final fixes #1

+++ Controllers/AdminController.cs
@@ -15,74 +15,89 @@ namespace ThrowdownAttire.Controllers
         [HttpPost]
         public ActionResult Create(ShirtCreateViewModel model)
         {
-            if ((bool)Session["Authenticated"] == true)
+            if (!Authenticated())
             {
-                if(model.title == null || Globals.Shirts.Exists(x => x.Title.ToLower() == model.title.ToLower()))
-                {
-                    return RedirectToAction("Admin", "Auth", new RouteValueDictionary(
-                        new {success = 0,  message = "Create shirt failed: please provide a unique title." }));
-                }
-
-                var repo = new ShirtRepository();
-
-                try {
-                    var displayBool = bool.Parse(model.display);
-                    if (model.newseries != null)
-                    {
-                        model.series = model.newseries;
-                        var url = repo.uploadSeriesImage(model.series, model.sliderimage);
-                        url = displayBool ? url : null;
-
-                        Globals.Types.Add(model.series, url);
-                    }
-
-                    var shirt = repo.Create(model);
-                    Globals.Shirts.Add(shirt);
-                    Globals.Shirts.Sort((x, y) => x.Title.CompareTo(y.Title));
-                    Globals.Types = Globals.Types.OrderBy(x => x.Key.Length).ToDictionary(x => x.Key, x => x.Value);
-                }
-                catch(Exception e)
-                {
-                    return RedirectToAction("Admin", "Auth", new RouteValueDictionary(
-                        new { success = 0, message = "Create shirt failed: " + e.Message }));
-                }
-
-                return RedirectToAction("Admin", "Auth", new RouteValueDictionary(
-                        new { success = 1, message = model.title + " successfully added to database."}));
+                return RedirectToAction("AdminLogin", "Auth");
             }
-            return RedirectToAction("AdminLogin", "Auth");
+            if (model.title == null || Globals.Shirts.Exists(x => x.Title.ToLower() == model.title.ToLower()))
+            {
+                return RedirectToAction("Admin", "Auth", new RouteValueDictionary(
+                    new { success = 0, message = "Create shirt failed: please provide a unique title." }));
+            }
+
+            var repo = new ShirtRepository();
+
+            try
+            {
+                var displayBool = bool.Parse(model.display);
+                if (model.newseries != null)
+                {
+                    model.series = model.newseries;
+                    var url = repo.uploadSeriesImage(model.series, model.sliderimage);
+                    url = displayBool ? url : null;
+
+                    Globals.Types.Add(model.series, url);
+                }
+
+                var shirt = repo.Create(model);
+                Globals.Shirts.Add(shirt);
+                Globals.Shirts.Sort((x, y) => x.Title.CompareTo(y.Title));
+                Globals.Types = Globals.Types.OrderBy(x => x.Key.Length).ToDictionary(x => x.Key, x => x.Value);
+            }
+            catch (Exception e)
+            {
+                return RedirectToAction("Admin", "Auth", new RouteValueDictionary(
+                    new { success = 0, message = "Create shirt failed: " + e.Message }));
+            }
+
+            return RedirectToAction("Admin", "Auth", new RouteValueDictionary(
+                    new { success = 1, message = model.title + " successfully added to database." }));
+
         }
 
         [HttpGet]
         public ActionResult Edit(string Id)
         {
-            if ((bool)Session["Authenticated"] == true)
+            if (!Authenticated())
             {
-                return View(Globals.Shirts.FirstOrDefault( x => x.Id.ToString() == Id));
+                return RedirectToAction("AdminLogin", "Auth");
             }
-            return RedirectToAction("AdminLogin", "Auth");
+            return View(Globals.Shirts.FirstOrDefault(x => x.Id.ToString() == Id));
         }
 
         [HttpPost]
         public ActionResult Edit(ShirtEditViewModel model)
         {
-            if ((bool)Session["Authenticated"] == true)
+            if (!Authenticated())
             {
-                var repo = new ShirtRepository();
-                var shirtDoc = repo.UpdateShirt(model);
-                var shirt = repo.createShirtFromBson(shirtDoc);
-
-                var i = Globals.Shirts.FindIndex(x => x.Id == shirt.Id);
-                Globals.Shirts[i] = shirt;
-
-                return RedirectToAction("Admin", "Auth");
+                return RedirectToAction("AdminLogin", "Auth");
             }
-            return RedirectToAction("AdminLogin", "Auth");
+
+            var repo = new ShirtRepository();
+            var shirtDoc = repo.UpdateShirt(model);
+            var shirt = repo.createShirtFromBson(shirtDoc);
+            var oldShirt = repo.FindShirtById(shirt.Id.ToString());
+
+            var i = Globals.Shirts.FindIndex(x => x.Id == shirt.Id);
+            Globals.Shirts[i] = shirt;
+
+            if (repo.firstShirtOfType(oldShirt.Type) == null)
+            {
+                Globals.Types.Remove(oldShirt.Type);
+                repo.deleteImage(oldShirt.Type);
+            }
+
+            return RedirectToAction("Admin", "Auth");
         }
 
         [HttpPost]
         public ActionResult UploadPhoto(HttpPostedFileBase[] images, string id)
         {
+            if (!Authenticated())
+            {
+                return RedirectToAction("AdminLogin", "Auth");
+            }
+
             var repo = new ShirtRepository();
             var shirt = repo.FindShirtById(id);
 
@@ -108,10 +123,15 @@ namespace ThrowdownAttire.Controllers
         [HttpPost]
         public ActionResult DeletePhoto(string src, string id)
         {
+            if (!Authenticated())
+            {
+                return RedirectToAction("AdminLogin", "Auth");
+            }
+
             var repo = new ShirtRepository();
             var shirt = repo.FindShirtById(id);
 
-            var pub_id = src.Split('/').Last().Replace(".jpg", "");
+            var pub_id = src.Split('/').Last().Replace(".png", "");
 
             try
             {
@@ -134,19 +154,22 @@ namespace ThrowdownAttire.Controllers
         [HttpPost]
         public ActionResult DeleteShirt(string id)
         {
-            var repo = new ShirtRepository();
-            repo.deleteShirt(id);
+            if (!Authenticated())
+            {
+                return RedirectToAction("AdminLogin", "Auth");
+            }
 
-            var shirt = repo.FindShirtById(id);
+            var repo = new ShirtRepository();
+            var shirt = repo.deleteShirt(id);
 
             Globals.Shirts.Remove(shirt);
 
-            if(Globals.Shirts.FirstOrDefault(x => x.Type == shirt.Type) == null)
+            if (repo.firstShirtOfType(shirt.Type) == null)
             {
                 Globals.Types.Remove(shirt.Type);
                 repo.deleteImage(shirt.Type);
             }
-            else if(Globals.Shirts.FirstOrDefault(x => x.Type == shirt.Type && x.Display == true) == null)
+            else if (repo.firstDisplayedShirtOfType(shirt.Type) == null)
             {
                 Globals.Types[shirt.Type] = null;
             }
@@ -157,18 +180,42 @@ namespace ThrowdownAttire.Controllers
         [HttpPost]
         public ActionResult SetDisplay(string id, string display)
         {
-            var repo = new ShirtRepository();
-            repo.UpdateShirt(id, "display", display);
-
-            var shirt = repo.FindShirtById(id);
-            shirt.Display = bool.Parse(display);
-
-            if(Globals.Shirts.FirstOrDefault(x => x.Type == shirt.Type && x.Display == true) == null)
+            if (!Authenticated())
             {
-                Globals.Types[shirt.Type] = null;
+                return RedirectToAction("AdminLogin", "Auth");
             }
 
+            var repo = new ShirtRepository();
+
+            repo.UpdateShirt(id, "display", display);
+            var shirt = repo.FindShirtById(id);
+
+            shirt.Display = bool.Parse(display);
+
+            Globals.Types[shirt.Type] = 
+                repo.firstDisplayedShirtOfType(shirt.Type) == null ? 
+                null : "http://res.cloudinary.com/throw-down-attire/image/upload/" + shirt.Type + ".png";
+
             return RedirectToAction("Admin", "Auth");
+        }
+
+        [HttpPost]
+        public ActionResult UpdateSlider(string type, HttpPostedFileBase image)
+        {
+            if (!Authenticated())
+            {
+                return RedirectToAction("AdminLogin", "Auth");
+            }
+
+            var repo = new ShirtRepository();
+            var src = repo.updateSeriesImage(type, image);
+
+            return new JsonResult() { Data = new { src = src } };
+        }
+
+        private bool Authenticated()
+        {
+            return (bool)Session["Authenticated"];
         }
     }
 }
